Started implementation of XML-based components
