+++ app/src/main/res/layout/activity_main.xml
@@ -68,7 +68,7 @@
         android:layout_gravity="start"
         android:choiceMode="singleChoice"
         android:divider="@android:color/darker_gray"
-        android:dividerHeight="1dp"
+        android:dividerHeight="0dp"
         android:background="@android:color/white"/>
